Refactor: React component files capitalized
Follow standards for React components
